# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Growth OS API", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/session": {
      "get": {
        "operationId": "session",
        "responses": {
          "200": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": { "type": "string" },
                    "workspaceId": { "type": "string" }
                  },
                  "required": ["userId", "workspaceId"],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_FAILED",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "RESOURCE_NOT_FOUND",
                            "INVALID_TRANSITION",
                            "IDEMPOTENCY_KEY_REUSED",
                            "RATE_LIMITED",
                            "CONFLICT",
                            "INTERNAL_ERROR",
                            "CSRF_FAILED",
                            "DUPLICATE_CUSTOMER",
                            "DUPLICATE_REVIEW_REQUIRED",
                            "CONSENT_REQUIRED",
                            "CAMPAIGN_VERSION_CONFLICT",
                            "CURRENCY_LOCKED",
                            "ACCOUNT_DELETION_REQUIRES_CONFIRMATION"
                          ]
                        },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" },
                        "fieldErrors": {
                          "type": "object",
                          "additionalProperties": { "type": "array", "items": { "type": "string" } }
                        },
                        "details": { "type": "object", "additionalProperties": {} }
                      },
                      "required": ["code", "message", "requestId"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspace": {
      "get": {
        "operationId": "getWorkspace",
        "responses": {
          "200": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "businessName": { "type": "string" },
                    "category": { "nullable": true, "type": "string" },
                    "timezone": { "type": "string" },
                    "currency": { "type": "string" },
                    "defaultCountryCode": { "type": "string" },
                    "bookingLink": { "nullable": true, "type": "string" },
                    "followUpDays": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "onboardingComplete": { "type": "boolean" },
                    "reviewResponseTemplate": { "type": "string" }
                  },
                  "required": [
                    "id",
                    "businessName",
                    "category",
                    "timezone",
                    "currency",
                    "defaultCountryCode",
                    "bookingLink",
                    "followUpDays",
                    "onboardingComplete",
                    "reviewResponseTemplate"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_FAILED",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "RESOURCE_NOT_FOUND",
                            "INVALID_TRANSITION",
                            "IDEMPOTENCY_KEY_REUSED",
                            "RATE_LIMITED",
                            "CONFLICT",
                            "INTERNAL_ERROR",
                            "CSRF_FAILED",
                            "DUPLICATE_CUSTOMER",
                            "DUPLICATE_REVIEW_REQUIRED",
                            "CONSENT_REQUIRED",
                            "CAMPAIGN_VERSION_CONFLICT",
                            "CURRENCY_LOCKED",
                            "ACCOUNT_DELETION_REQUIRES_CONFIRMATION"
                          ]
                        },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" },
                        "fieldErrors": {
                          "type": "object",
                          "additionalProperties": { "type": "array", "items": { "type": "string" } }
                        },
                        "details": { "type": "object", "additionalProperties": {} }
                      },
                      "required": ["code", "message", "requestId"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateWorkspace",
        "responses": {
          "200": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "businessName": { "type": "string" },
                    "category": { "nullable": true, "type": "string" },
                    "timezone": { "type": "string" },
                    "currency": { "type": "string" },
                    "defaultCountryCode": { "type": "string" },
                    "bookingLink": { "nullable": true, "type": "string" },
                    "followUpDays": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "onboardingComplete": { "type": "boolean" },
                    "reviewResponseTemplate": { "type": "string" }
                  },
                  "required": [
                    "id",
                    "businessName",
                    "category",
                    "timezone",
                    "currency",
                    "defaultCountryCode",
                    "bookingLink",
                    "followUpDays",
                    "onboardingComplete",
                    "reviewResponseTemplate"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_FAILED",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "RESOURCE_NOT_FOUND",
                            "INVALID_TRANSITION",
                            "IDEMPOTENCY_KEY_REUSED",
                            "RATE_LIMITED",
                            "CONFLICT",
                            "INTERNAL_ERROR",
                            "CSRF_FAILED",
                            "DUPLICATE_CUSTOMER",
                            "DUPLICATE_REVIEW_REQUIRED",
                            "CONSENT_REQUIRED",
                            "CAMPAIGN_VERSION_CONFLICT",
                            "CURRENCY_LOCKED",
                            "ACCOUNT_DELETION_REQUIRES_CONFIRMATION"
                          ]
                        },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" },
                        "fieldErrors": {
                          "type": "object",
                          "additionalProperties": { "type": "array", "items": { "type": "string" } }
                        },
                        "details": { "type": "object", "additionalProperties": {} }
                      },
                      "required": ["code", "message", "requestId"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "businessName": { "type": "string", "minLength": 2, "maxLength": 120 },
                  "category": {
                    "nullable": true,
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 80
                  },
                  "timezone": { "type": "string", "minLength": 1, "maxLength": 100 },
                  "currency": { "type": "string", "pattern": "^[A-Z]{3}$" },
                  "defaultCountryCode": { "type": "string", "pattern": "^\\+[1-9][0-9]{0,3}$" },
                  "bookingLink": {
                    "anyOf": [
                      {
                        "anyOf": [
                          { "type": "string", "maxLength": 500, "format": "uri" },
                          { "type": "string", "enum": [""] }
                        ]
                      },
                      { "type": "string", "nullable": true, "enum": [null] }
                    ]
                  },
                  "followUpDays": { "type": "integer", "minimum": 1, "maximum": 90 },
                  "reviewResponseTemplate": { "type": "string", "maxLength": 2000 },
                  "onboardingComplete": { "type": "boolean" }
                },
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
    "/api/v1/workspace/account": {
      "delete": {
        "operationId": "deleteWorkspaceAccount",
        "responses": {
          "204": { "description": "No content" },
          "400": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_FAILED",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "RESOURCE_NOT_FOUND",
                            "INVALID_TRANSITION",
                            "IDEMPOTENCY_KEY_REUSED",
                            "RATE_LIMITED",
                            "CONFLICT",
                            "INTERNAL_ERROR",
                            "CSRF_FAILED",
                            "DUPLICATE_CUSTOMER",
                            "DUPLICATE_REVIEW_REQUIRED",
                            "CONSENT_REQUIRED",
                            "CAMPAIGN_VERSION_CONFLICT",
                            "CURRENCY_LOCKED",
                            "ACCOUNT_DELETION_REQUIRES_CONFIRMATION"
                          ]
                        },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" },
                        "fieldErrors": {
                          "type": "object",
                          "additionalProperties": { "type": "array", "items": { "type": "string" } }
                        },
                        "details": { "type": "object", "additionalProperties": {} }
                      },
                      "required": ["code", "message", "requestId"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_FAILED",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "RESOURCE_NOT_FOUND",
                            "INVALID_TRANSITION",
                            "IDEMPOTENCY_KEY_REUSED",
                            "RATE_LIMITED",
                            "CONFLICT",
                            "INTERNAL_ERROR",
                            "CSRF_FAILED",
                            "DUPLICATE_CUSTOMER",
                            "DUPLICATE_REVIEW_REQUIRED",
                            "CONSENT_REQUIRED",
                            "CAMPAIGN_VERSION_CONFLICT",
                            "CURRENCY_LOCKED",
                            "ACCOUNT_DELETION_REQUIRES_CONFIRMATION"
                          ]
                        },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" },
                        "fieldErrors": {
                          "type": "object",
                          "additionalProperties": { "type": "array", "items": { "type": "string" } }
                        },
                        "details": { "type": "object", "additionalProperties": {} }
                      },
                      "required": ["code", "message", "requestId"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": { "type": "string", "minLength": 1, "maxLength": 128 },
                  "businessNameConfirmation": { "type": "string", "minLength": 2, "maxLength": 120 }
                },
                "required": ["password", "businessNameConfirmation"],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
  }
}